Target size in EnsembleMember
I tried to make it default but it could break your code
Here's what's changed:

To have a common `train_epoch` for arb. ensemble member.
We need to know the target size, this can not be inferred from
the output size since we do not know the modelling distr. (can be
whatever)

Previous solution was a hack I did for regression which now in the pale
cast of thought does not seem at all good for classification.